Fix a bug that distributed tensors/embeddings are destroyed unexpectedly. (#27)
*Issue #, if available:*
When we have a deep copy of a model object, it has a deep copy of
DistTensor and DistEmbedding objects. When such a local object is
destroyed, the data in the remote server will be destroyed as well. To
fix the issue, we should avoid deep copy of these objects locally when a
model is deep copied.

*Description of changes:*
Avoid deep copy of DistTensor and DistEmbedding objects.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Da Zheng <zhengda@192.168.0.100>
Co-authored-by: Ubuntu <ubuntu@ip-172-31-10-98.us-west-2.compute.internal>
Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>