Support node data in separate files. (#104)
*Description of changes:*
In some cases, the node data of the same node type are stored in
separate files. We need to use multiple node blocks to specify the
files. Currently, we only support this for node data.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-15-191.us-west-2.compute.internal>