Increase `eval_batch_size` to accelerate lp inference (#66)
Resolves issue #81 

This PR updates the default `eval_batch_size` to 10K, replacing the
previous value of 1024 as commonly set in most `.yaml` files. The
increase in batch size from 1K to 10K results in a significant reduction
in end-to-end inference time, from 45 minutes to 19 minutes. In order to
enhance the usability of the default value of 10K, the eval_batch_size
flag is removed from the `.yaml` files. Users who encounter
out-of-memory (OOM) issues can easily adjust the `eval_batch_size` to a
lower value as needed.

Following table has the timings for link prediction on OGBN-papers100M
(number of nodes = 111M, number of edges = 1.6B). The dataset is
partitioned into 3. g4dn.metal instances are used.

#### Environment setup: 
* CUDA 11.3
* DGL 1.0.1
* Pytorch 1.12.0

#### Hyperparameter:
* model_encoder_type: rgcn
* n_layers: 1, 
* n_hidden: 16, 
* fanout: "20", 
* batch_size: 128


| eval_batch_size | Time (min) lp_mini_batch_predict | Time (min) full
inference pipeline |
| ---------------- | ---------------------------------- |
-------------------------------- |
| 1K  | 31  | 45 |
| 10K  | 5 | 19 |
| 100K  | 2 | 14 |

---------

Co-authored-by: Israt Nisa <nisisrat@amazon.com>
Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>