Optimize external memory operation on HDF5 in graph construction (#122)
*Description of changes:*
HDF5 cannot support slicing data on disks. We need to convert all data
in HDF5 to Numpy arrays. If the node/edge data are stored in multiple
HDF5 files, we need to merge them into one Numpy array. The Numpy array
may be backed by SSDs.
Another optimization is that if we don't need to perform any
transformation on the data in HDF5 files, we don't need to use
multiprocessing to load the data. Instead, we can just use the master
process to copy the data to Numpy array directly.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-15-191.us-west-2.compute.internal>