Add multi-target node training support (#338)
*Issue #, if available:*

*Description of changes:*

Previous GraphStorm implementation assume only a single target node type
is used to compute loss and gradient. In this version, we the loss and
gradient can be computed on multiple node types.

However, for evaluation, we still only evaluate the model performance on
a single node type (the node type with the largest number of nodes for
now). To better support multiple target node type evaluation, we might
need to re-design our evaluator to better store the best score and its
associated model.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.