Use the neighbors to compute the features of featureless nodes. (#237)
*Description of changes:*
Currently, the default way of handling featureless nodes is to use
learnable embeddings. However, it's very slow when the model has
learnable embeddings. This PR is to use the neighbor nodes to compute
the features of the featureless nodes before running the GNN model.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-12-121.us-west-2.compute.internal>