Refactor the code of BERT computation to fix a potential bug. (#426)
*Description of changes:*
The original code locates a BERT model based on the index. There is
potentially a bug here. If a user fine-tune BERT models on two node
types:
```
  node_lm_models:
    -
      lm_type: bert
      model_name: "bert-base-uncased"
      gradient_checkpoint: true
      node_types:
        - paper
    -
      lm_type: bert
      model_name: "bert-base-uncased"
      gradient_checkpoint: true
      node_types:
        - fos
```
If the user trains the GNN model with the following config,
```
  node_lm_models:
    -
      lm_type: bert
      model_name: "bert-base-uncased"
      gradient_checkpoint: true
      node_types:
        - fos
    -
      lm_type: bert
      model_name: "bert-base-uncased"
      gradient_checkpoint: true
      node_types:
        - paper
```
The original code will use the wrong BERT model for the node types.
The fix is to associate the BERT model with node types and the parameter
names contain the node types to make sure the parameters are loaded to
the right BERT model.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-10-98.us-west-2.compute.internal>
Co-authored-by: Ubuntu <ubuntu@ip-172-31-12-121.us-west-2.compute.internal>