Fix the problem of caching twice (#493)
*Description of changes:*
In the GNN inference with BERT models, we first cache the BERT
embeddings and then cache the embeddings from the input layer. This
causes a lot of unnecessary memory consumption. An easy to avoid this is
to remove caching the BERT embeddings in the inference script.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Ubuntu <ubuntu@ip-172-31-12-121.us-west-2.compute.internal>