Change save prediction logic of node prediction tasks (#543)
Originally, the saving node prediction result process saves the
prediction results of all the nodes of target node type(s) regardless of
whether a node belongs to an inference set or not. The same thing
happends for the saving node embedding process. In this PR, we only save
the prediction results and node embeddings of target nodes in the
inference set.

Third PR of #508
Also Fix: #519

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Xiang Song <xiangsx@amazon.com>