[Feature] Add gat_encoder (#602)
*Issue #, if available:*

*Description of changes:*
This PR adds a GATEncoder and a test case in test_gnn. 

This PR does not handle 0-in-degree case. As DGL warning, this may lead
to invalid results.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-35-48.us-west-2.compute.internal>