Fix bug in constructing node features from neighbor nodes with only BERT embeddings (#561)
*Description of changes:*
When constructing node features from neighbor nodes, the sampler reads
the node features from the dataset to identify the right neighbor nodes
to compute node features. If the features of the neighbor nodes are text
features, the current datasets don't know these features. This PR
basically changes this behavior. When the BERT embeddings are cached,
they are added to the datasets as node features. In this way, the
sampler for constructing node features can identify the right neighbor
nodes.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-12-121.us-west-2.compute.internal>
Co-authored-by: Ubuntu <ubuntu@ip-172-31-0-17.us-west-2.compute.internal>
Co-authored-by: Ubuntu <ubuntu@ip-172-31-7-202.us-west-2.compute.internal>