[GSProcessing] Allow re-partitioning to run on the Spark leader (#767)
*Issue #, if available:*

*Description of changes:*

With this change we allow the re-partition step to run on the leader
instance, removing the need to run a separate job for it. If no files
need re-partitioning, and only Parquet modification is needed for
labels/masks we _always_ apply that on the leader.

To accomplish this we make the following changes:

* Add new `--do-repartition` argument
* Refactor repartition_files.py to allow separate execution of just
Parquet metadata modification, and full repartition that runs both the
row count alignment and the Parquet metadata modification.
* As part of this refactoring we also split the previously verbose
`main` function into multiple smaller functions to allow easier
maintenance. Each step is now performed by a separate function, namely
`_repartition_edge_files`, `_repartition_node_files` and
`modify_flat_array_metadata`. A combined public function
`repartition_files` calls these three functions as needed in sequence.
All functions take the graph metadata dict as input, and a
ParquetRepartitioner instance.
* Introduce checks for node data and Parquet metadata modification in
our unit tests. Test coverage for `repartition_files.py` is now
increased to 89% (from 81%).

## Testing

Unit testing, and ran jobs on SageMaker and EMR-S with ml25M. For
large-scale random 100M-edges-10M-nodes graph, the repartition step
completed successfully on SageMaker (large memory instances) and failed
(gracefully) on EMR-S. When the repart step fails we print an error in
the logs to indicate to the user they'll need to run the additional
gs-repartition job separately.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>
Co-authored-by: jalencato <jalencato23pistons@gmail.com>