[Docker] Move the Amazon ECR Public Gallery auth to only the CPU branch of the local build script. (#832)
*Issue #, if available:*

*Description of changes:*
* To pull the CPU source image we need to authenticate with Amazon
Public ECR gallery using the AWS cli. This is not needed to build the
GPU image, so we move that auth step under the CPU branch, to avoid
customers having to add the permission to their roles if they're only
building the GPU image.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.