[Doc] Fix the type error in example Notebook 2 (#886)
*Issue #, if available:*

*Description of changes:*
This PR fix the type error of `reverse_edge_types_map` in the API
example notebook 2. When it should be a dictionary. When set the
`exclude_training_targets` to be `True`, this map will be used.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Ubuntu <ubuntu@ip-172-31-35-48.us-west-2.compute.internal>