[GraphBolt] Add GraphBolt graph construction for GConstruct and GPartition (GraphBolt PR2) (#1011)
*Issue #, if available:*

*Description of changes:*

* Add `--use-graphbolt` argument to GConstruct and GSPartition entry
points. When set to `true`, we convert the graph data to GraphBolt
format after partitioning.
* Add new `gpartition_test.sh` set of integration tests for GSPartition
entry point.
* Add new
tests/end2end-tests/graphbolt-gs-integration/graphbolt-graph-construction.sh
set of integration tests for graph construction with GraphBolt enabled.
Requires DGL >= 2.x.
* To accommodate this, we we modify
`tests/end2end-tests/data_gen/process_movielens.py` so that it also
produces a JSON file that can be used as input the the DistPart DGL
pipeline.
* To parse and handle different DGL versions we use `importlib.metadata`
and the `packaging` modules. We introduce an explicit dependency on
`packaging` to GSF. This shouldn't change the user environment, as we
already had a transitive dependency on `packaging` through the
`transformers` lib.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>