[Doc] API Notebook example for Graph Prediction Solution (#1026)
*Issue #, if available:*

*Description of changes:*
This PR is the 2nd PR of Graph Prediction Solution example. This PR
provides a Notebook to explain how to customize a GNN encoder that uses
`super-node` method for graph prediction solution. The Notebook use the
OGBG data generated in the graph prediction example.

Preview of rendered Notebook:
https://james4graphstorm.readthedocs.io/en/james_nb5/api/notebooks/Notebook_5_GP_solution.html

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-33-210.us-west-2.compute.internal>
Co-authored-by: EC2 Default User <ec2-user@ip-172-16-9-169.us-west-2.compute.internal>
Co-authored-by: Theodore Vasiloudis <theodoros.vasiloudis@gmail.com>