[Doc] Refactor configuration doc and release config.get_argument_parser API (#1028)
*Issue #, if available:*

*Description of changes:*
This PR is the 1st one for the Notebook 6 PRs. This PR revises the model
training and inference configuration document, moving the `--cf`
argument out of launch argument set, and set it separately. In addition,
this PR releases GraphStorm config module's `get_argument_parser` method
for the Notebook 6, which will convert customized models that use
GraphStorm APIs to use CLIs.

The preview pages include:
1. New configurations page:
https://james4graphstorm.readthedocs.io/en/james_nb6_docs/cli/model-training-inference/configuration-run.html
2. New API index page:
https://james4graphstorm.readthedocs.io/en/james_nb6_docs/api/references/index.html
3. Config API pages:
https://james4graphstorm.readthedocs.io/en/james_nb6_docs/api/references/graphstorm.config.html
and
https://james4graphstorm.readthedocs.io/en/james_nb6_docs/api/generated/graphstorm.config.get_argument_parser.html#graphstorm.config.get_argument_parser

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-33-210.us-west-2.compute.internal>
Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>
Co-authored-by: jalencato <jalencato23pistons@gmail.com>