[GraphBolt] Add integration tests for LP and NC training and inference. (#1029)
*Issue #, if available:*

*Description of changes:*

* Add a test script that executes LP and NC training and inference tasks
with GraphBolt enabled. To run on one instance we create-single
partition versions of the ml-100k data, and use the same gconstruct/yaml
configs as we do in our end2end CPU tests.
* Merging will close https://github.com/awslabs/graphstorm/issues/982

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.