[SageMaker] [GraphBolt] Add support for launching GraphBolt jobs on SageMaker (#1083)
*Issue #, if available:*


*Description of changes:*

* Add a new SageMaker job to convert DistPart data to GraphBolt. This is
our only option currently as there's no way to directly use S3 as a
writable, shared file system in SageMaker, see
https://github.com/awslabs/graphstorm/issues/1081 for details.
* The `sagemaker/launch_graphbolt_convert.py` will launch the SageMaker
job, that downloads the entire partitioned graph to one instance, then
runs the GB conversion, one partition at a time. Because DGL writes the
new fused CSC graph representation in the same directory as the input
data, we can't use one of SageMaker's FastFile modes to stream the data,
as that creates read-only filesystems.
* [Optional] We also include an example of how one could use a SageMaker
Pipeline to run the GSPartition and GBConvert jobs in sequence, but this
can be removed (because SageMaker Pipelines are persistent once
created).
* Added unit test mechanism to test sagemaker scripts, we start with
testing our parsing logic. To make the scripts available to the runner's
python runtime we add the `graphstorm/sagemaker/launch` directory to the
runner's `PYTHONPATH`.

EDIT: One note about the PR: The changes to the partition launch that
use a SageMaker Pipeline are for demonstration purposes, I think I'll
remove them alltogether and just have separate partition/gbconvert jobs.
But we might want to have an example of how to programmatically build an
SM pipeline as an example, e.g. from gsprocessing to training (as SM
jobs)







By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>