[GSProcessing] Fix parsing of file paths that start with "./" (#1091)
*Issue #, if available:*

*Description of changes:*

* Some GConstruct or GSProcessing config files from customers could have
file paths that start with `./` which works when input is local but
break for S3. We fix this by removing the `./` prefix if it exists when
parsing the input files.
* Add change to config we use for testing to verify the change works
(although we'd probably need a more S3-specific test)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.