[Bug Fix] Fix Edge Feature None block error (#1128)
*Issue #, if available:*

*Description of changes:*

Previously (before adding edge feature in message passing), we do not
require any block information, so the code works fine. Now as we require
them for edge feats, we need to make a small fix about it. The error is
only triggered in large dataset with workers > 24. Already test
successfully with amazon-review dataset with this PR.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.