[GSProcessing] Enforce re-order for node label processing during classification (#1136)
*Issue #, if available:*

Fixes #1135 
Fixes #1138

*Description of changes:*

* We guarantee ordering for node label classification by ordering the
transformed label DF after processing by the NODE_INT_MAPPING id, and
doing the same for masks.
* Because there is no guarantee for order when writing to Parquet from
Spark _even for ordered Spark DataFrames_, we collect the labels and
masks to a Pandas DF on the Spark leader, and write that using pyarrow.
* NOTE: DistPart requires at least N files to be present for every
node/edge type, where N is the number of requested partitions. So when
writing back the masks and label we still use the number of incoming
partitions to create as many files.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.