[GSProcessing] Make label transformation use Spark for output instead of Pyarrow when using custom masks. (#1150)
*Issue #, if available:*

*Description of changes:*

* Previously, during classification labels would be collected to the
driver and written using pyarrow regardless of the type of mask that
will be generated.
* To improve scaling, since we observed that the collect is not
necessary when using custom masks, we revert to writing the label files
using Spark, avoiding collecting to the driver and improving
scalability.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: jalencato <jalencato23pistons@gmail.com>