[SageMaker] Add launcher and support for SageMaker HPO jobs (#1133)
*Issue #, if available:*

Fixes https://github.com/awslabs/graphstorm/issues/1072

*Description of changes:*

* We change train_entry and sagmaker_train so that when the user does
not provide a value for the `--model-artifact-s3` argument, we rely on
the SageMaker service to upload model artifacts. This makes it possible
to run HPO and other jobs that rely on using the "official" SageMaker
model paths.
* Add example launch script for HPO jobs and documentation.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>