Fix the error case when the label is a 1D tensor for regression tasks (#1199)
*Issue #, if available:*
#1198 

*Description of changes:*
The prediction results of regression tasks are always 2D tensors, even
when the prediction target is a single float value for each node. If the
label is a 1D tensor, it will trigger a tensor shape mismatch error.
This PR fixes this error.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Xiang Song <xiangsx@amazon.com>