[Fix] Fix an HGT Encoder Corner Case: 0 edge in a block (#1196)
*Issue #, if available:*

*Description of changes:*
This PR fixes an HGT Encoder corner case, where the edge type with
feature has 0 edge in a block. In such case the input of edge feature
become empty dict, and will trigger an assertion error in the previous
implementation. This PR handles this corner case.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-0-244.us-west-2.compute.internal>
Co-authored-by: RonaldBXu <72748153+RonaldBXu@users.noreply.github.com>
Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>