Apply suggestions for Noop transformation (#1228)
*Issue #, if available:*

*Description of changes:*

* Apply final comments from #1225 
* Also change test names for other transformations. Previously all
transformations were tested under a single unit test. This meant that if
one of the tests failed the following ones would not run which is not
the intended behavior when testing, we don't want tests to depend on
each other. This also allows the individual tests to be discoverable and
testable within an IDE.
* Also fixes a pylint error that was introduced in #1227 


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.