[Ehancement] Refactor `return_proba` warning message and document (#1236)
*Issue #, if available:*

*Description of changes:*
This PR comment out the warning message of `set return_proba to be
true`, which is confusing during training pipeline. And revise the
document of `return_proba` configuration to explain this configuration
only works on inference output.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-11-9.us-west-2.compute.internal>
Co-authored-by: Theodore Vasiloudis <theodoros.vasiloudis@gmail.com>