[GSPartition] Ensure random partition assigns at least one node per partition (#1239)
*Issue #, if available:*

*Description of changes:*

* Previously it was possible for random partition to create assignments
that led to some node types having 0 nodes in one partition. This would
eventually trigger errors in DistPart.
* Similarly, we allowed the creation of partitions for node types with
fewer nodes that the requested partitions.
* This PR ensures every partition will have at least one node, and
raises an error if we try to create a partition for a node type with
fewer nodes than the requested partitions.
* Expand tests to cover all cases.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>
Co-authored-by: Jian Zhang (James) <6593865@qq.com>
Co-authored-by: Ubuntu <ubuntu@ip-172-31-11-9.us-west-2.compute.internal>