[Feature] Pack model artifact func for realtime inference (#1251)
*Issue #, if available:*

*Description of changes:*
This PR add a packing function and a script to pack trained model
artifacts into a `model.tar.gz` file. This function will be a part of
launch script for realtime inference deployment.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-11-9.us-west-2.compute.internal>
Co-authored-by: xiang song(charlie.song) <classicxsong@gmail.com>
Co-authored-by: Theodore Vasiloudis <theodoros.vasiloudis@gmail.com>