Add support for Knowledge graph embedding training. (#1260)
*Issue #, if available:*

*Description of changes:*

We introduce a GSPureLearnableInputLayer that it create learnable
embeddings for each node in the graph. With GSPureLearnableInputLayer,
we can train a KGE model (i.e., learnable embeddings + relation
embeddings only)




By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Xiang Song <xiangsx@amazon.com>