[Bug]Fix the inference code for non-autoregressive mode (#1281)
*Issue #, if available:*

*Description of changes:*
This PR fix a problem in the inferrer, where non-autoregressive mode
does not work in the previous version. Now add the `use-ar` argument to
the inferrer, and check if use autoregressive.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

Co-authored-by: Ubuntu <ubuntu@ip-172-31-11-9.us-west-2.compute.internal>