[Bug] Set logging default level in real-time launch script (#1317)
*Issue #, if available:*

*Description of changes:*
This PR fix the error of no log output information in the launch script.
By this PR, we set the default logging level to INFO, which will output
both endpoint ARN and name.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-4-143.us-west-2.compute.internal>
Co-authored-by: Theodore Vasiloudis <theodoros.vasiloudis@gmail.com>