[sagemaker] [rt-inference] Add option to deploy real-time endpoint into VPC, allow custom entry point location (#1326)
*Issue #, if available:*

* To use NeptuneDB as a graph source, the endpoint needs to be able to
communicate with the NDB instance, which is always placed inside a VPC.
* With this PR we add the ability to select VPC subnets to place the
endpoint in, and additional security groups that allow it to communicate
with other resources in the VPC, e.g. an inference client that acts as
an intermediate between NDB and SM endpoint, or directly communicate
with NDB.
* We also add an option to provide the location of the entry point to
use when launching the endpoint. This allows custom entry point
locations to be used (e.g. we use this when using the launch script in
isolation on a SageMaker notebook instance)

*Description of changes:*

* Add two new optional arguments to
`sagemaker/launch/launch_realtime_endpoint.py`, `--vpc-subnet-ids` and
`--vpc-security-group-ids`, that allow us to define which VPC to deploy
the endpoint in, and which security groups to attach to it.
* Add new optional `--entry-point-path` argument to allow user to define
which entry point to use when launching the endpoint.
* Update docs with new args

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.